# client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Colegio Api",
    "contact": {
      "name": "Leidy Tatiana Sanchez ",
      "email": "ltsa0314@gmail.com"
    },
    "version": "v1"
  },
  "paths": {
    "/api/AlumnoAsignaturas": {
      "get": {
        "tags": [
          "AlumnoAsignaturas"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "AlumnoAsignaturas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoAsignaturaEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoAsignaturaEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoAsignaturaEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "AlumnoAsignaturas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoAsignaturaEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoAsignaturaEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoAsignaturaEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AlumnoAsignaturas/{alumnoId}/{asignaturaId}/{ano}": {
      "get": {
        "tags": [
          "AlumnoAsignaturas"
        ],
        "parameters": [
          {
            "name": "alumnoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "asignaturaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ano",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/AlumnoAsignaturas/{id}": {
      "delete": {
        "tags": [
          "AlumnoAsignaturas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Alumnos": {
      "get": {
        "tags": [
          "Alumnos"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Alumnos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Alumnos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnoEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Alumnos/Report": {
      "get": {
        "tags": [
          "Alumnos"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Alumnos/{id}": {
      "get": {
        "tags": [
          "Alumnos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Alumnos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Asignaturas": {
      "get": {
        "tags": [
          "Asignaturas"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Asignaturas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AsignaturaEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AsignaturaEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AsignaturaEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Asignaturas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AsignaturaEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AsignaturaEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AsignaturaEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Asignaturas/{id}": {
      "get": {
        "tags": [
          "Asignaturas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Asignaturas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ProfesorAsignaturas": {
      "get": {
        "tags": [
          "ProfesorAsignaturas"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ProfesorAsignaturas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorAsignaturaEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorAsignaturaEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorAsignaturaEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "ProfesorAsignaturas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorAsignaturaEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorAsignaturaEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorAsignaturaEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ProfesorAsignaturas/{id}": {
      "get": {
        "tags": [
          "ProfesorAsignaturas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ProfesorAsignaturas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profesores": {
      "get": {
        "tags": [
          "Profesores"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Profesores"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Profesores"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorEntity"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorEntity"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProfesorEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profesores/{id}": {
      "get": {
        "tags": [
          "Profesores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Profesores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProfesorEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "identificacion": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          },
          "edad": {
            "type": "integer",
            "format": "int32"
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "integer",
            "format": "int32"
          },
          "asignatura": {
            "$ref": "#/components/schemas/ProfesorAsignaturaEntity"
          }
        },
        "additionalProperties": false
      },
      "ProfesorAsignaturaEntity": {
        "type": "object",
        "properties": {
          "profesorId": {
            "type": "integer",
            "format": "int32"
          },
          "asignaturaId": {
            "type": "integer",
            "format": "int32"
          },
          "profesor": {
            "$ref": "#/components/schemas/ProfesorEntity"
          },
          "asignatura": {
            "$ref": "#/components/schemas/AsignaturaEntity"
          }
        },
        "additionalProperties": false
      },
      "AsignaturaEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "alumnoAsignaturas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlumnoAsignaturaEntity"
            },
            "nullable": true
          },
          "profesorAsignatura": {
            "$ref": "#/components/schemas/ProfesorAsignaturaEntity"
          }
        },
        "additionalProperties": false
      },
      "AlumnoEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "identificacion": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          },
          "edad": {
            "type": "integer",
            "format": "int32"
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "integer",
            "format": "int32"
          },
          "alumnoAsignaturas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlumnoAsignaturaEntity"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AlumnoAsignaturaEntity": {
        "type": "object",
        "properties": {
          "alumnoId": {
            "type": "integer",
            "format": "int32"
          },
          "asignaturaId": {
            "type": "integer",
            "format": "int32"
          },
          "calificacionFinal": {
            "type": "number",
            "format": "double"
          },
          "ano": {
            "type": "integer",
            "format": "int32"
          },
          "asignatura": {
            "$ref": "#/components/schemas/AsignaturaEntity"
          },
          "alumno": {
            "$ref": "#/components/schemas/AlumnoEntity"
          }
        },
        "additionalProperties": false
      }
    }
  }
}